# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cavendo Engine API
  description: |
    REST API for Cavendo Engine - an AI agent task management platform.

    ## Authentication

    The API supports three authentication methods:

    ### Session Authentication (Cookie-based)
    Used by the web UI. Login via `/api/auth/login` to receive a session cookie.

    ### Agent Key Authentication (X-Agent-Key header)
    Used by AI agents. Generate keys via the admin UI or `/api/agents/{id}/keys`.
    Keys start with `cav_ak_` prefix.

    ### User Key Authentication (X-Agent-Key header)
    Personal API keys for users (e.g., for MCP integration). Generate via `/api/users/me/keys`.
    Keys start with `cav_uk_` prefix.

    ## Key Concepts

    - **Agents**: AI agents that execute tasks (Claude, GPT, custom implementations)
    - **Tasks**: Work items assigned to agents
    - **Deliverables**: Work output submitted by agents for review
    - **Projects**: Organize tasks and knowledge
    - **Knowledge**: Project-scoped reference documents for agents
    - **Webhooks**: Event notifications for agents

  version: 0.1.0
  contact:
    name: Cavendo Engine
    url: https://github.com/Cavendo/Engine
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://your-server.com
    description: Production
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie obtained from login endpoint
    agentKey:
      type: apiKey
      in: header
      name: X-Agent-Key
      description: Agent API key (cav_ak_...) or User API key (cav_uk_...)
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            errors:
              type: object
              additionalProperties:
                type: string
...
    Agent:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Research Agent
        type:
          type: string
          enum: [autonomous, semi-autonomous, supervised]
          example: semi-autonomous
        description:
          type: string
          nullable: true
        capabilities:
          type: array
          items:
            type: string
          example: [research, writing, analysis]
        status:
          type: string
          enum: [active, inactive, suspended]
          example: active
        webhook_url:
          type: string
          format: uri
          nullable: true
        webhook_events:
          type: array
          items:
            type: string
        max_concurrent_tasks:
          type: integer
          minimum: 1
          maximum: 100
          example: 3
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentWithKeys:
      allOf:
        - $ref: '#/components/schemas/Agent'
        - type: object
          properties:
            keys:
              type: array
              items:
                $ref: '#/components/schemas/AgentKey'
    AgentKey:
      type: object
      properties:
        id:
          type: integer
        key_prefix:
          type: string
          example: cav_ak_a
        name:
          type: string
          nullable: true
        scopes:
          type: array
          items:
            type: string
            enum: [read, write, webhook:create, '*']
        last_used_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    UpdateAgentRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        type:
          type: string
          enum: [autonomous, semi-autonomous, supervised]
        description:
          type: string
          maxLength: 1000
          nullable: true
        capabilities:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
        status:
          type: string
          enum: [active, inactive, suspended]
        webhookUrl:
          type: string
          format: uri
          nullable: true
        webhookEvents:
          type: array
          items:
            type: string
        maxConcurrentTasks:
          type: integer
          minimum: 1
          maximum: 100
...
    AgentActivity:
      type: object
      properties:
        id:
          type: integer
        agent_id:
          type: integer
        action:
          type: string
          example: deliverable.submitted
        resource_type:
          type: string
          nullable: true
          example: deliverable
        resource_id:
          type: integer
          nullable: true
        details:
          type: object
          additionalProperties: true
        ip_address:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        agent_name:
          type: string
          description: Populated when querying with joins
    ActivityStats:
      type: object
      properties:
        period:
          type: string
          enum: ['24h', '7d', '30d']
        totalActions:
          type: integer
        actionsByType:
          type: array
          items:
            type: object
            properties:
              action:
                type: string
              count:
                type: integer
        actionsByAgent:
          type: array
          items:
            type: object
            properties:
              agent_id:
                type: integer
              agent_name:
                type: string
              count:
                type: integer
        actionsOverTime:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              count:
                type: integer

    # User Key schemas
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Session required
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Insufficient permissions
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
paths:
...
  /api/agents/{id}:
    get:
      tags: [Agents]
      summary: Get agent details
      description: Returns agent details including API keys.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AgentWithKeys'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Agents]
      summary: Update agent
      description: Updates agent configuration. Requires admin role.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Agent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Agents]
      summary: Delete agent
      description: Deletes an agent and all associated keys. Requires admin role.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Agent deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
                        example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/activity:
    get:
      tags: [Activity]
      summary: List agent activity
      description: Returns agent activity log with optional filtering.
      security:
        - sessionAuth: []
      parameters:
        - name: agentId
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: resourceType
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Activity log
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentActivity'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/activity/stats:
    get:
      tags: [Activity]
      summary: Get activity statistics
      description: Returns aggregated activity statistics for a time period.
      security:
        - sessionAuth: []
      parameters:
        - name: agentId
          in: query
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
            enum: ['24h', '7d', '30d']
            default: '7d'
      responses:
        '200':
          description: Activity statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ActivityStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/activity/agents/{id}:
    get:
      tags: [Activity]
      summary: Get agent activity
      description: Returns activity for a specific agent with summary statistics.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Agent activity
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      agent:
                        type: object
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                      stats:
                        type: object
                        properties:
                          total_actions:
                            type: integer
                          unique_actions:
                            type: integer
                          first_activity:
                            type: string
                            format: date-time
                          last_activity:
                            type: string
                            format: date-time
                      activities:
                        type: array
                        items:
                          $ref: '#/components/schemas/AgentActivity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # User endpoints
...